chore(storage-to-table): Reorganized components structure + refactoring
