+++ src/components/data-workflows/storage-to-table/configuration/StorageToTableConfOverview.vue
