+++ src/components/data-workflows/storage-to-table/run/GcsToGbqRuns.vue
@@ -129,7 +129,7 @@ import store from '@/store';
 import moment from 'moment';
 import _ from 'lodash';
 import Util from '@/util';
-import DataManagementFilters from '../../common/DataManagementFilters';
+import DataManagementFilters from '../../../common/DataManagementFilters';
 
 export default {
 	components: {
