+++ src/components/data-workflows/storage-to-table/run/StorageToTableRunView.vue
