+++ src/util/data-workflows/index.js
@@ -3,7 +3,9 @@ export const getWorkflowName = name => {
 		gbqToGcsConfs: 'GBQ to GCS',
 		GbqToGcsConf: 'GBQ to GCS',
 		TableToStorageConf: 'Table to Storage',
-		TableToStorageConfs: 'Table to Storage'
+		TableToStorageConfs: 'Table to Storage',
+		GcsToGbqRuns: 'Storage to Table',
+		GcsToGbqConfs: 'Storage to Table'
 	};
 
 	return items[name];
