--- src/views/data-workflows/gcs-to-gbq/StorageToTableConf.vue
@@ -1,28 +1,0 @@
-<template>
-	<div>
-		<DataManagementHeader :tabsItems="tabsItems" :headerTitle="headerTitle" />
-		<StorageToTableConf />
-	</div>
-</template>
-
-<script>
-import StorageToTableConf from '../../../components/data-workflows/storage-to-table/StorageToTableConf';
-import DataManagementHeader from '../../../components/app/headers/DataManagementHeader';
-
-export default {
-	components: {
-		StorageToTableConf,
-		DataManagementHeader
-	},
-	data() {
-		return {
-			headerTitle: 'Storage to Table',
-			tabsItems: [
-				{ id: 1, title: 'Runs', link: '/storagetotable/runs' },
-				{ id: 2, title: 'Configurations', link: '/storagetotable/confs' },
-				{ id: 3, title: 'Documentation', link: '' }
-			]
-		};
-	}
-};
-</script>
