+++ src/views/data-workflows/storage-to-table/tab-items.js
@@ -1,0 +1,11 @@
+export default {
+	computed: {
+		tabItems() {
+			return [
+				{ id: 1, title: 'Runs', link: '/storagetotable/runs' },
+				{ id: 2, title: 'Configurations', link: '/storagetotable/confs' },
+				{ id: 3, title: 'Documentation', link: '' }
+			];
+		}
+	}
+};
